pemula_machine_learning_klasifikasi_gambar
The class is intended for beginners who want to start their career in machine learning with reference to industry competency standards. At the end of the class, students can learn the basic material and algorithms of Machine Learning, and be able to implement them in creating Machine Learning models to process data.